refactor: move the gc activities to a dedicated file
