+++ connectors/src/connectors/zendesk/temporal/activities.ts
@@ -18,6 +18,7 @@ import {
   fetchZendeskArticlesInCategory,
   fetchZendeskTicketsInBrand,
 } from "@connectors/connectors/zendesk/lib/zendesk_api";
+import { ZENDESK_BATCH_SIZE } from "@connectors/connectors/zendesk/temporal/config";
 import { dataSourceConfigFromConnector } from "@connectors/lib/api/data_source_config";
 import { concurrentExecutor } from "@connectors/lib/async_utils";
 import { deleteFromDataSource } from "@connectors/lib/data_sources";
@@ -33,8 +34,6 @@ import {
   ZendeskTicketResource,
 } from "@connectors/resources/zendesk_resources";
 import type { DataSourceConfig } from "@connectors/types/data_source_config";
-
-const ZENDESK_BATCH_SIZE = 100;
 
 /**
  * This activity is responsible for updating the lastSyncStartTime of the connector to now.
@@ -184,56 +183,6 @@ export async function syncZendeskBrandActivity({
     helpCenterAllowed: brandInDb.helpCenterPermission === "read",
     ticketsAllowed: brandInDb.ticketsPermission === "read",
   };
-}
-
-/**
- * This activity is responsible for fetching a batch of tickets
- * that are older than the retention period and ready to be deleted.
- */
-export async function getNextOldTicketBatchActivity(
-  connectorId: ModelId
-): Promise<number[]> {
-  const configuration =
-    await ZendeskConfigurationResource.fetchByConnectorId(connectorId);
-  if (!configuration) {
-    throw new Error(
-      `[Zendesk] Configuration not found, connectorId: ${connectorId}`
-    );
-  }
-  return ZendeskTicketResource.fetchOutdatedTicketIds({
-    connectorId,
-    expirationDate: new Date(
-      Date.now() - configuration.retentionPeriodDays * 24 * 60 * 60 * 1000 // conversion from days to ms
-    ),
-    batchSize: ZENDESK_BATCH_SIZE,
-  });
-}
-
-/**
- * This activity is responsible for deleting a batch of tickets given their IDs.
- */
-export async function deleteTicketBatchActivity(
-  connectorId: ModelId,
-  ticketIds: number[]
-): Promise<void> {
-  const connector = await ConnectorResource.fetchById(connectorId);
-  if (!connector) {
-    throw new Error("[Zendesk] Connector not found.");
-  }
-  const dataSourceConfig = dataSourceConfigFromConnector(connector);
-  const loggerArgs = {
-    workspaceId: dataSourceConfig.workspaceId,
-    connectorId,
-    provider: "zendesk",
-    dataSourceId: dataSourceConfig.dataSourceId,
-  };
-
-  await concurrentExecutor(
-    ticketIds,
-    (ticketId) =>
-      deleteTicket({ connectorId, ticketId, dataSourceConfig, loggerArgs }),
-    { concurrency: 10 }
-  );
 }
 
 /**
