+++ connectors/src/connectors/zendesk/temporal/config.ts
@@ -1,2 +1,4 @@
 export const WORKFLOW_VERSION = 4;
 export const QUEUE_NAME = `zendesk-queue-v${WORKFLOW_VERSION}`;
+
+export const ZENDESK_BATCH_SIZE = 100;
